feat: move state up the chain
