+++ packages/site/src/components/TokenForm.js
@@ -1,38 +1,27 @@
-import {
-	BrowserRouter as Router,
-	Switch,
-	Route,
-	useRouteMatch,
-	Link
-  } from "react-router-dom";
-
 import React, { Fragment } from "react";
 import { TokensList } from "./TokensList";
 import { JsonOutput } from "./JsonOutput";
 
-export const TokenForm = ( { brand, tokens } ) => {
-	let match = useRouteMatch();
+export const TokenForm = ( { action, tokens, brandName } ) => {
+
+	console.log(tokens);
 
 	return (
 		<Fragment>
-			<Route path={`${match.path}/:action`}>
-
-				<div className="lrv-a-grid lrv-a-cols3">
-					<section className="lrv-a-span2">
-						<TokensList
-							tokenType="Brand"
-							selectedBrand={brand}
-							tokens={tokens}
-						/>
-					</section>
-					<section className="lrv-u-width-100p lrv-u-overflow-auto">
-						<JsonOutput
-							tokensJson={tokens}
-						/>
-					</section>
-				</div>
-
-			</Route>
+			<div className="lrv-a-grid lrv-a-cols3">
+				<section className="lrv-a-span2">
+					<TokensList
+						action={action}
+						brandName={brandName}
+						tokens={tokens}
+					/>
+				</section>
+				<section className="lrv-u-width-100p lrv-u-overflow-auto">
+					<JsonOutput
+						tokensJson={tokens}
+					/>
+				</section>
+			</div>
 		</Fragment>
 	)
 };
